Remove PromQL based telemetry to avoid Tldr.
